backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
ultralytics
requests
pydantic
